requirements.txt
llama-index
python-dotenv
chromadb
PyPDF2
pathlib
asyncio
sentence-transformers